fix: 修复其他路径转换问题
