+++ internal/translator/codex/openai/chat-completions/codex_openai_request.go
@@ -10,6 +10,7 @@ import (
 	"strconv"
 	"strings"
 
+	"github.com/router-for-me/CLIProxyAPI/v7/internal/util"
 	"github.com/tidwall/gjson"
 	"github.com/tidwall/sjson"
 )
@@ -313,7 +314,7 @@ func ConvertOpenAIRequestToCodex(modelName string, inputRawJSON []byte, stream b
 						item, _ = sjson.SetBytes(item, "description", v.Value())
 					}
 					if v := fn.Get("parameters"); v.Exists() {
-						item, _ = sjson.SetRawBytes(item, "parameters", normalizeCodexToolSchema(v.Raw))
+						item, _ = sjson.SetRawBytes(item, "parameters", util.NormalizeCodexToolSchema(v.Raw))
 					}
 					if v := fn.Get("strict"); v.Exists() {
 						item, _ = sjson.SetBytes(item, "strict", v.Value())
@@ -357,71 +358,6 @@ func ConvertOpenAIRequestToCodex(modelName string, inputRawJSON []byte, stream b
 
 	out, _ = sjson.SetBytes(out, "store", false)
 	return out
-}
-
-func normalizeCodexToolSchema(raw string) []byte {
-	raw = strings.TrimSpace(raw)
-	if raw == "" || raw == "null" || !gjson.Valid(raw) {
-		return []byte(`{"type":"object","properties":{}}`)
-	}
-
-	schema := []byte(raw)
-	for _, path := range codexSchemaNullArrayFieldPaths(raw) {
-		schema, _ = sjson.DeleteBytes(schema, path)
-	}
-	return schema
-}
-
-func codexSchemaNullArrayFieldPaths(raw string) []string {
-	arrayFields := map[string]struct{}{
-		"allOf":    {},
-		"anyOf":    {},
-		"enum":     {},
-		"oneOf":    {},
-		"required": {},
-	}
-
-	var paths []string
-	walkCodexSchemaNullArrayFields(gjson.Parse(raw), "", arrayFields, &paths)
-	return paths
-}
-
-func walkCodexSchemaNullArrayFields(value gjson.Result, path string, arrayFields map[string]struct{}, paths *[]string) {
-	if value.IsArray() {
-		items := value.Array()
-		for i, item := range items {
-			walkCodexSchemaNullArrayFields(item, joinCodexSchemaPath(path, strconv.Itoa(i)), arrayFields, paths)
-		}
-		return
-	}
-	if !value.IsObject() {
-		return
-	}
-
-	value.ForEach(func(key, child gjson.Result) bool {
-		keyString := key.String()
-		childPath := joinCodexSchemaPath(path, escapeCodexSchemaPathKey(keyString))
-		if _, ok := arrayFields[keyString]; ok && child.Type == gjson.Null {
-			*paths = append(*paths, childPath)
-			return true
-		}
-		walkCodexSchemaNullArrayFields(child, childPath, arrayFields, paths)
-		return true
-	})
-}
-
-func joinCodexSchemaPath(base, suffix string) string {
-	if base == "" {
-		return suffix
-	}
-	return base + "." + suffix
-}
-
-func escapeCodexSchemaPathKey(key string) string {
-	if !strings.ContainsAny(key, ".*?") {
-		return key
-	}
-	return strings.NewReplacer(".", "\\.", "*", "\\*", "?", "\\?").Replace(key)
 }
 
 func setToolCallOutputContent(funcOutput []byte, content gjson.Result) []byte {
